Working basic chord detection with FFTW3.
Currently the client can pass in a frame size and sample rate for the chord detector but they are only placeholders to maybe implement later. The frame size is currently automatically set to arco's block size BL = 32 and sample rate is arco's AR = 44100.

Adding the chord detection text to the test messed up the indentation for the onset text, which I'm not sure how to fix.