Chord Detection: added confidence measure, half diminished chords detection capability, and slightly modified the amount of harmonics to reduce.
Confidence is calculated with softmax on chord scores then entropy on the probabilities.

Restructured ChordDetector.cpp code for easier testing and added a function for comparison between different levels of harmonics reduction.

Also added a `bool shouldDisplay` member to ChordDetector, to tell the client if the detected chord is confident enough to be displayed.